Make code more efficient, loop playerSelection
I was able to make the condition much more efficient by grouping the win conditions together instead of using switch;
I was also able to make the resulting strings considerably more efficient by using template literals.
One of the template literals was beginning a sentence in the returned string, so I needed to write a capitalize function.
I also made the playerSelection prompt loop when the user's input was invalid.